Moved a lot of the logic that was in my App.jsx file to my TodoContainer.jsx file to make it leaner, as well as to make sure that my fetch request is found in the correct file that the assignment from lesson 5.1 expects it to be in.
